Moved controller function for deadlines out to a new controller
