+++ src/main/twirl/views/manage/opportunityOverview.scala.html
@@ -70,14 +70,14 @@ backLink=Some(BackLink("Opportunity library", controllers.manage.routes.Opportun
                                 <tr>
                                     <td>
                                         2. <a id="op-setup-link2"
-                                              href="@controllers.manage.routes.OpportunityController.viewDeadlines(opportunity.id)"
+                                              href="@controllers.manage.routes.DeadlineController.view(opportunity.id)"
                                             > Opportunity deadlines</a>
                                     </td>
                                 </tr>
                                 <tr>
                                     <td>
                                         3. <a id="op-setup-link3"
-                                              href="@controllers.manage.routes.GrantValueController.viewGrantValue(opportunity.id)"
+                                              href="@controllers.manage.routes.GrantValueController.view(opportunity.id)"
                                             > Maximum grant value</a>
                                     </td>
                                 </tr>
